+++ lib/index.ts
@@ -1,8 +1,9 @@
 import { validate, prevalidate } from './validate';
 import { validateProfile } from './validate/profiles';
 import { readValue } from './validate/rows';
-import { getErrorLevel, getLabel } from './utils/helpers';
+import { getErrorLevel, getLabel, ErrorLevelEnum } from './utils/helpers';
 import profiles from './schema/profiles';
+import { PositionTypeEnum } from './schema/shema.type';
 
 export {
   validate,
@@ -12,21 +13,29 @@ export {
   readValue,
   getErrorLevel,
   profiles,
+  ErrorLevelEnum,
+  PositionTypeEnum,
 };
 
-export { ErrorLevelEnum } from './validate/validate.type';
+export type { ParseFileType } from './validate/parse/parse.type';
 
 export type {
-  PrevalidateType,
-  ValidateProfileType,
+  ProfilesValidationType,
   FieldType,
-  NotFoundFieldType,
-  ValidateRowType,
   ValueIsValidType,
   ValidateFileType,
-  ProfilesValidationType,
-  ParseFileType,
+  PrevalidateType,
   ProfileErrorType,
+  ErrorLevelType,
+  ValidateRowFullType,
+  NotFoundFieldLevelType,
+  ValidateType,
 } from './validate/validate.type';
+
+export type {
+  ParsedValue,
+  ReadValueType,
+  ParsedValues,
+} from './schema/shema.type';
 
 export type { ProfileType } from './schema/profiles/profile.type';
