+++ lib/utils/helpers.ts
@@ -4,7 +4,12 @@ import profiles from '../schema/profiles';
 import errorLabels from '../schema/error-labels';
 import Schema from '../schema/index';
 import { ProfileType } from '../schema/profiles/profile.type';
-import { ErrorLevelEnum } from '../validate/validate.type';
+
+export enum ErrorLevelEnum {
+  ERROR = 'E',
+  WARNING = 'W',
+  INFO = 'I',
+}
 
 // On fait une liste des langues régional que l'on peut utiliser
 export const allowedRegionalLangs: string[] = languesRegionales.map(
