+++ lib/validate/parse/parse.type.ts
@@ -1,0 +1,15 @@
+import { ParseError, ParseResult } from 'papaparse';
+
+export type ParseReturn = ParseResult<Record<string, string>> & {
+  encoding: string;
+};
+
+export type ParseFileType = {
+  encoding: string;
+  linebreak: string;
+  delimiter: string;
+  originalFields: string[];
+  parseOk: boolean;
+  parseErrors: ParseError[];
+  parsedRows: Record<string, string>[];
+};
