+++ lib/validate/validate.type.ts
@@ -1,91 +1,14 @@
-import { ParseError, ParseResult } from 'papaparse';
+import { ParsedValues } from '../schema/shema.type';
+import { ErrorLevelEnum } from '../utils/helpers';
+import { ParseFileType } from './parse/parse.type';
+
 export const IS_TOPO_NB = '99999';
 
-export type ParseReturn = ParseResult<Record<string, string>> & {
-  encoding: string;
-};
+// VALIDATE ROW
 
-export type ParseFileType = {
-  encoding: string;
-  linebreak: string;
-  delimiter: string;
-  originalFields: string[];
-  parseOk: boolean;
-  parseErrors: ParseError[];
-  parsedRows: Record<string, string>[];
-};
-
-export enum ErrorLevelEnum {
-  ERROR = 'E',
-  WARNING = 'W',
-  INFO = 'I',
-}
-
-export type ProfileErrorType = {
+export type ErrorType = {
   code: string;
-  level: ErrorLevelEnum;
-};
-
-export type ParsedValue =
-  | string
-  | string[]
-  | boolean
-  | number
-  | PositionTypeEnum
-  | Date
-  | undefined;
-
-export type ReadValueType = {
-  parsedValue: ParsedValue;
-  additionalValues: any;
-  errors: string[];
-};
-
-export type ValidateRowFullType = ValidateRowType & { isValid: boolean };
-
-export enum PositionTypeEnum {
-  ENTREE = 'entrée',
-  BATIMENT = 'bâtiment',
-  CAGE_ESCALIER = 'cage d’escalier',
-  LOGEMENT = 'logement',
-  SERVICE_TECHNIQUE = 'service technique',
-  DELIVRANCE_POSTALE = 'délivrance postale',
-  PARCELLE = 'parcelle',
-  SEGMENT = 'segment',
-}
-
-export type CommuneNomIsoCodeKey = `commune_nom_${string}`;
-export type CommuneDelegueeNomIsoCodeKey = `commune_deleguee_nom_${string}`;
-export type VoieNomIsoCodeKey = `voie_nom_${string}`;
-export type LieuditComplementNomIsoCodeKey = `lieudit_complement_nom_${string}`;
-
-export type ParsedValues = {
-  uid_adresse?: string;
-  id_ban_commune?: string;
-  id_ban_toponyme?: string;
-  id_ban_adresse?: string;
-  cle_interop?: string;
-  commune_insee?: string;
-  commune_nom?: string;
-  commune_deleguee_insee?: string;
-  commune_deleguee_nom?: string;
-  voie_nom?: string;
-  lieudit_complement_nom?: string;
-  numero?: number;
-  suffixe?: string;
-  position?: PositionTypeEnum;
-  x?: number;
-  y?: number;
-  long?: number;
-  lat?: number;
-  cad_parcelles?: string[];
-  source?: string;
-  date_der_maj?: Date;
-  certification_commune?: boolean;
-  [key: CommuneNomIsoCodeKey]: string;
-  [key: CommuneDelegueeNomIsoCodeKey]: string;
-  [key: VoieNomIsoCodeKey]: string;
-  [key: LieuditComplementNomIsoCodeKey]: string;
+  schemaName?: string;
 };
 
 export type ValidateRowType = {
@@ -93,9 +16,11 @@ export type ValidateRowType = {
   parsedValues: ParsedValues;
   additionalValues: Record<string, any>;
   localizedValues: Record<string, any>;
-  errors: { code: string; schemaName?: string; level?: ErrorLevelEnum }[];
+  errors: ErrorType[];
   line: number;
 };
+
+// PREVALIDATE
 
 export type ProfilesValidationType = {
   code: string;
@@ -137,6 +62,28 @@ export type PrevalidateType = Omit<ParseFileType, 'parsedRows'> & {
   uniqueErrors: string[];
 };
 
-export type ValidateProfileType = PrevalidateType & {
+// VALIDATE PROFILE
+
+export type ProfileErrorType = {
+  code: string;
+  level: ErrorLevelEnum;
+};
+
+export type ErrorLevelType = ErrorType & {
+  level?: ErrorLevelEnum;
+};
+
+export type ValidateRowFullType = Omit<ValidateRowType, 'errors'> & {
+  isValid: boolean;
+  errors: ErrorLevelType[];
+};
+
+export type NotFoundFieldLevelType = NotFoundFieldType & {
+  level?: ErrorLevelEnum;
+};
+
+export type ValidateType = Omit<PrevalidateType, 'notFoundFields' | 'rows'> & {
   profilErrors: ProfileErrorType[];
+  notFoundFields: NotFoundFieldLevelType[];
+  rows: ValidateRowFullType[];
 };
